feat(non_profit): enhance TOR and deployment system with user-based filters, location tracking, and methodology improvements
- Updated Terms of Reference (TOR) and Deployment Request filters to use the logged-in User and linked Employee records.
- Ensures users can only view TORs and deployments within their organizational hierarchy.
- Refactored deployment structure by renaming Personnel Deployment Assignment to Request for clearer semantics.
- Introduced Sub County and Ward DocTypes for granular location tracking and better geographic data management.
- Added resource cost calculations to Personnel Terms of Reference for improved financial tracking.
- Enhanced TOR Approach Methods with updated field types, labels, and better clarity.
- Improved Project form filters based on expected end date and linked company descendants.
- Added utility function to retrieve company descendants for hierarchical filtering.
- Updated organization type management to improve structural hierarchy handling.
- Added test cases for new functionalities (Sub County, Ward, TOR Methodology).

This update strengthens the deployment and TOR systems by introducing user-aware filtering, enhanced location granularity, and refined methodology tracking — ensuring more accurate, contextual, and secure data management.
