+++ non_profit/fixtures/notification.json
@@ -15,7 +15,7 @@
     "message_type": "Markdown",
     "method": "after_insert",
     "modified": "2025-09-02 11:40:15.773877",
-    "module": null,
+    "module": "Non Profit",
     "name": "Volunteer Signup",
     "print_format": null,
     "property_value": null,
