+++ non_profit/fixtures/organisation_type.json
@@ -1,0 +1,20 @@
+[
+ {
+  "docstatus": 0,
+  "doctype": "Organisation Type",
+  "modified": "2025-10-24 13:50:50.723943",
+  "name": "Region"
+ },
+ {
+  "docstatus": 0,
+  "doctype": "Organisation Type",
+  "modified": "2025-10-24 13:50:59.876746",
+  "name": "Branch"
+ },
+ {
+  "docstatus": 0,
+  "doctype": "Organisation Type",
+  "modified": "2025-10-24 13:51:09.424405",
+  "name": "Company"
+ }
+]
