+++ non_profit/non_profit/custom/job_applicant.json
@@ -809,7 +809,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:45:56.999125",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-column_break_0n1h6",
       "no_copy": 0,
       "non_negative": 0,
@@ -1129,7 +1129,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:42:17.423032",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-column_break_uhlvo",
       "no_copy": 0,
       "non_negative": 0,
@@ -1449,7 +1449,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-08 02:05:50.715998",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-opportunity_type",
       "no_copy": 0,
       "non_negative": 0,
@@ -1833,7 +1833,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:42:17.574884",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-eligibility_status",
       "no_copy": 0,
       "non_negative": 0,
@@ -3305,7 +3305,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:38:26.988538",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-screening_assessment_tab",
       "no_copy": 0,
       "non_negative": 0,
@@ -3369,7 +3369,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:39:15.157421",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-screening_question_responses",
       "no_copy": 0,
       "non_negative": 0,
@@ -3433,7 +3433,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:45:56.870551",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-screening_score_percent",
       "no_copy": 0,
       "non_negative": 0,
@@ -3497,7 +3497,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:42:17.675836",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-section_break_fur99",
       "no_copy": 0,
       "non_negative": 0,
@@ -3881,7 +3881,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 05:42:17.296849",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-total_score",
       "no_copy": 0,
       "non_negative": 0,
@@ -4226,7 +4226,7 @@
       "is_system_generated": 0,
       "modified": "2025-10-08 02:05:50.616760",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Applicant-main-field_order",
       "owner": "Administrator",
       "property": "field_order",
