+++ non_profit/non_profit/custom/job_opening.json
@@ -489,7 +489,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 09:26:14.064038",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Opening-profession",
       "no_copy": 0,
       "non_negative": 0,
@@ -809,7 +809,7 @@
       "mandatory_depends_on": null,
       "modified": "2025-10-07 06:40:18.398837",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Opening-job_opening_template",
       "no_copy": 0,
       "non_negative": 0,
@@ -1876,7 +1876,7 @@
       "is_system_generated": 0,
       "modified": "2025-10-07 09:26:13.993973",
       "modified_by": "Administrator",
-      "module": null,
+      "module": "Non Profit",
       "name": "Job Opening-main-field_order",
       "owner": "Administrator",
       "property": "field_order",
