+++ non_profit/non_profit/custom/material_request.json
@@ -163,7 +163,7 @@
       "insert_after": "deployment",
       "is_system_generated": 0,
       "is_virtual": 0,
-      "label": "Personnel Deployment Assignment",
+      "label": "Personnel Deployment Request",
       "length": 0,
       "link_filters": null,
       "mandatory_depends_on": null,
@@ -173,7 +173,7 @@
       "name": "Material Request-personnel_deployment_assignment",
       "no_copy": 0,
       "non_negative": 0,
-      "options": "Personnel Deployment Assignment",
+      "options": "Personnel Deployment Request",
       "owner": "Administrator",
       "permlevel": 0,
       "placeholder": null,
@@ -227,7 +227,7 @@
       "insert_after": "column_break_zktkg",
       "is_system_generated": 0,
       "is_virtual": 0,
-      "label": "Personnel Deployment Request",
+      "label": "Deployment Request Tool",
       "length": 0,
       "link_filters": null,
       "mandatory_depends_on": null,
@@ -237,7 +237,7 @@
       "name": "Material Request-personnel_deployment_request",
       "no_copy": 0,
       "non_negative": 0,
-      "options": "Personnel Deployment Request",
+      "options": "Deployment Request Tool",
       "owner": "Administrator",
       "permlevel": 0,
       "placeholder": null,
