+++ non_profit/non_profit/doctype/administrative_location_table/administrative_location_table.js
@@ -1,0 +1,8 @@
+// Copyright (c) 2025, Frappe and contributors
+// For license information, please see license.txt
+
+// frappe.ui.form.on("Administrative Location Table", {
+// 	refresh(frm) {
+
+// 	},
+// });
