+++ non_profit/non_profit/doctype/administrative_location_table/administrative_location_table.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2025, Frappe and contributors
+# For license information, please see license.txt
+
+# import frappe
+from frappe.model.document import Document
+
+
+class AdministrativeLocationTable(Document):
+	pass
