+++ non_profit/non_profit/doctype/administrative_location_table/test_administrative_location_table.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestPersonnelDeploymentRequest(FrappeTestCase):
+class TestAdministrativeLocationTable(FrappeTestCase):
 	pass
