+++ non_profit/non_profit/doctype/deployment_request_tool/test_deployment_request_tool.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestPersonnelDeploymentRequest(FrappeTestCase):
-	pass
+class TestDeploymentRequestTool(FrappeTestCase):
+    pass
