+++ non_profit/non_profit/doctype/organisation_type/test_organisation_type.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestPersonnelDeploymentRequest(FrappeTestCase):
+class TestOrganisationType(FrappeTestCase):
 	pass
