--- non_profit/non_profit/doctype/personnel_deployment_assignment/personnel_deployment_assignment.json
@@ -1,337 +1,0 @@
-{
- "actions": [],
- "allow_rename": 1,
- "autoname": "format:PDA-{YYYY}-{MM}-{####}",
- "creation": "2025-09-18 13:32:25.091012",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "details_tab",
-  "details_section",
-  "employee",
-  "employee_name",
-  "column_break_qbmp",
-  "project",
-  "task",
-  "location",
-  "column_break_qlbs",
-  "status",
-  "company",
-  "deployment",
-  "require_contract_before_deployment",
-  "dates__reason_section",
-  "expected_start_date",
-  "expected_end_date",
-  "column_break_hmmj",
-  "notes",
-  "terms_section",
-  "terms_of_reference",
-  "term_details",
-  "offer_terms",
-  "approval_section",
-  "deployment_approver",
-  "deployment_approver_name",
-  "amended_from",
-  "column_break_xpwb",
-  "date",
-  "expense_approver",
-  "advance_approver"
- ],
- "fields": [
-  {
-   "fieldname": "column_break_qbmp",
-   "fieldtype": "Column Break"
-  },
-  {
-   "allow_on_submit": 1,
-   "default": "Pending",
-   "fieldname": "status",
-   "fieldtype": "Select",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Status",
-   "options": "Pending\nAccepted\nRejected\nCancelled",
-   "reqd": 1
-  },
-  {
-   "fieldname": "employee",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Employee",
-   "options": "Employee",
-   "reqd": 1,
-   "search_index": 1
-  },
-  {
-   "bold": 1,
-   "fetch_from": "employee.first_name",
-   "fieldname": "employee_name",
-   "fieldtype": "Data",
-   "label": "Employee Name"
-  },
-  {
-   "fieldname": "details_section",
-   "fieldtype": "Section Break",
-   "label": "Details"
-  },
-  {
-   "fieldname": "project",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Project",
-   "options": "Project",
-   "reqd": 1,
-   "search_index": 1
-  },
-  {
-   "fieldname": "task",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Task",
-   "options": "Task"
-  },
-  {
-   "fieldname": "location",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Location",
-   "options": "Location"
-  },
-  {
-   "fieldname": "column_break_qlbs",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fetch_from": "employee.company",
-   "fieldname": "company",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Company",
-   "options": "Company",
-   "reqd": 1
-  },
-  {
-   "fieldname": "dates__reason_section",
-   "fieldtype": "Section Break",
-   "label": "Dates & Reason"
-  },
-  {
-   "fieldname": "expected_start_date",
-   "fieldtype": "Datetime",
-   "label": "Expected Start Date",
-   "reqd": 1
-  },
-  {
-   "fieldname": "expected_end_date",
-   "fieldtype": "Datetime",
-   "label": "Expected End Date"
-  },
-  {
-   "fieldname": "column_break_hmmj",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "notes",
-   "fieldtype": "Text Editor",
-   "label": "Notes"
-  },
-  {
-   "fieldname": "terms_section",
-   "fieldtype": "Section Break",
-   "label": "Terms"
-  },
-  {
-   "fieldname": "offer_terms",
-   "fieldtype": "Table",
-   "label": "Offer Terms",
-   "options": "Job Offer Term"
-  },
-  {
-   "fieldname": "approval_section",
-   "fieldtype": "Section Break",
-   "label": "Approval"
-  },
-  {
-   "allow_on_submit": 1,
-   "fieldname": "deployment_approver",
-   "fieldtype": "Link",
-   "label": "Deployment Approver",
-   "options": "User"
-  },
-  {
-   "fetch_from": "deployment_approver.full_name",
-   "fieldname": "deployment_approver_name",
-   "fieldtype": "Data",
-   "label": "Deployment Approver Name"
-  },
-  {
-   "fieldname": "column_break_xpwb",
-   "fieldtype": "Column Break"
-  },
-  {
-   "fieldname": "date",
-   "fieldtype": "Date",
-   "label": "Date"
-  },
-  {
-   "fieldname": "deployment",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_standard_filter": 1,
-   "label": "Deployment",
-   "options": "Personnel Deployment Request",
-   "search_index": 1
-  },
-  {
-   "fieldname": "amended_from",
-   "fieldtype": "Link",
-   "label": "Amended From",
-   "no_copy": 1,
-   "options": "Personnel Deployment Assignment",
-   "print_hide": 1,
-   "read_only": 1,
-   "search_index": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "require_contract_before_deployment",
-   "fieldtype": "Check",
-   "label": "Require Contract Before Deployment"
-  },
-  {
-   "fieldname": "terms_of_reference",
-   "fieldtype": "Link",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Terms of Reference",
-   "options": "Personnel Terms of Reference",
-   "reqd": 1,
-   "search_index": 1
-  },
-  {
-   "fieldname": "term_details",
-   "fieldtype": "Text Editor",
-   "hidden": 1,
-   "label": "Term Details"
-  },
-  {
-   "fieldname": "details_tab",
-   "fieldtype": "Tab Break",
-   "label": "Details",
-   "show_dashboard": 1
-  },
-  {
-   "allow_on_submit": 1,
-   "fieldname": "expense_approver",
-   "fieldtype": "Link",
-   "label": "Expense Approver",
-   "options": "User",
-   "reqd": 1
-  },
-  {
-   "allow_on_submit": 1,
-   "fieldname": "advance_approver",
-   "fieldtype": "Link",
-   "label": "Advance Approver",
-   "options": "User",
-   "reqd": 1
-  }
- ],
- "grid_page_length": 50,
- "index_web_pages_for_search": 1,
- "is_calendar_and_gantt": 1,
- "is_submittable": 1,
- "links": [
-  {
-   "group": "Support",
-   "link_doctype": "Issue",
-   "link_fieldname": "personnel_deployment_assignment"
-  },
-  {
-   "group": "Activity",
-   "link_doctype": "Timesheet",
-   "link_fieldname": "personnel_deployment_assignment"
-  },
-  {
-   "group": "Material",
-   "link_doctype": "Material Request",
-   "link_fieldname": "personnel_deployment_assignment"
-  },
-  {
-   "group": "Claims",
-   "link_doctype": "Expense Claim",
-   "link_fieldname": "personnel_deployment_assignment"
-  },
-  {
-   "group": "Claims",
-   "link_doctype": "Employee Advance",
-   "link_fieldname": "personnel_deployment_assignment"
-  },
-  {
-   "group": "Contract",
-   "link_doctype": "Contract",
-   "link_fieldname": "personnel_deployment_assignment"
-  }
- ],
- "modified": "2025-09-30 08:20:54.616184",
- "modified_by": "Administrator",
- "module": "Non Profit",
- "name": "Personnel Deployment Assignment",
- "naming_rule": "Expression",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "row_format": "Dynamic",
- "sort_field": "modified",
- "sort_order": "DESC",
- "states": [
-  {
-   "color": "Yellow",
-   "title": "Pending"
-  },
-  {
-   "color": "Green",
-   "title": "Accepted"
-  },
-  {
-   "color": "Red",
-   "title": "Rejected"
-  },
-  {
-   "color": "Orange",
-   "title": "Cancelled"
-  }
- ]
-}
