+++ non_profit/non_profit/doctype/sub_county_table/sub_county_table.json
@@ -1,0 +1,37 @@
+{
+ "actions": [],
+ "allow_rename": 1,
+ "creation": "2025-10-24 12:29:43.939789",
+ "default_view": "List",
+ "doctype": "DocType",
+ "editable_grid": 1,
+ "engine": "InnoDB",
+ "field_order": [
+  "sub_county"
+ ],
+ "fields": [
+  {
+   "fieldname": "sub_county",
+   "fieldtype": "Link",
+   "in_list_view": 1,
+   "label": "Sub County",
+   "options": "Sub County",
+   "reqd": 1
+  }
+ ],
+ "grid_page_length": 50,
+ "index_web_pages_for_search": 1,
+ "istable": 1,
+ "links": [],
+ "modified": "2025-10-24 14:37:17.440319",
+ "modified_by": "Administrator",
+ "module": "Non Profit",
+ "name": "Sub County Table",
+ "owner": "Administrator",
+ "permissions": [],
+ "row_format": "Dynamic",
+ "rows_threshold_for_grid_search": 20,
+ "sort_field": "modified",
+ "sort_order": "DESC",
+ "states": []
+}
