+++ non_profit/non_profit/doctype/tor_approach_methods/tor_approach_methods.json
@@ -5,20 +5,10 @@
  "doctype": "DocType",
  "engine": "InnoDB",
  "field_order": [
-  "method",
+  "methodology",
   "notes"
  ],
  "fields": [
-  {
-   "fieldname": "method",
-   "fieldtype": "Data",
-   "in_filter": 1,
-   "in_list_view": 1,
-   "in_preview": 1,
-   "in_standard_filter": 1,
-   "label": "Method",
-   "reqd": 1
-  },
   {
    "fieldname": "notes",
    "fieldtype": "Small Text",
@@ -28,13 +18,24 @@
    "in_standard_filter": 1,
    "label": "Notes",
    "reqd": 1
+  },
+  {
+   "fieldname": "methodology",
+   "fieldtype": "Link",
+   "in_filter": 1,
+   "in_list_view": 1,
+   "in_preview": 1,
+   "in_standard_filter": 1,
+   "label": "Methodology",
+   "options": "TOR Methodology",
+   "reqd": 1
   }
  ],
  "grid_page_length": 50,
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2025-10-23 13:10:32.461652",
+ "modified": "2025-10-24 11:55:07.951773",
  "modified_by": "Administrator",
  "module": "Non Profit",
  "name": "TOR Approach Methods",
