+++ non_profit/non_profit/doctype/tor_methodology/test_tor_methodology.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestPersonnelDeploymentAssignment(FrappeTestCase):
+class TestTORMethodology(FrappeTestCase):
 	pass
