+++ non_profit/non_profit/doctype/tor_methodology/tor_methodology.py
@@ -1,0 +1,9 @@
+# Copyright (c) 2025, Frappe and contributors
+# For license information, please see license.txt
+
+# import frappe
+from frappe.model.document import Document
+
+
+class TORMethodology(Document):
+	pass
