+++ non_profit/non_profit/doctype/tor_resources/tor_resources.json
@@ -44,14 +44,16 @@
   },
   {
    "fieldname": "donor",
-   "fieldtype": "Data",
+   "fieldtype": "Link",
    "in_filter": 1,
    "in_preview": 1,
-   "label": "Donor Code",
+   "label": "Donor",
+   "options": "Donor",
    "reqd": 1
   },
   {
    "columns": 1,
+   "default": "1",
    "fieldname": "quantity",
    "fieldtype": "Float",
    "in_filter": 1,
@@ -93,7 +95,7 @@
    "in_preview": 1,
    "in_standard_filter": 1,
    "label": "Total Cost",
-   "reqd": 1
+   "read_only": 1
   },
   {
    "fieldname": "project_code",
@@ -106,7 +108,7 @@
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2025-10-23 12:55:24.564604",
+ "modified": "2025-10-24 12:52:12.201895",
  "modified_by": "Administrator",
  "module": "Non Profit",
  "name": "TOR Resources",
