+++ non_profit/non_profit/doctype/ward/test_ward.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestPersonnelDeploymentRequest(FrappeTestCase):
+class TestWard(FrappeTestCase):
 	pass
