+++ non_profit/non_profit/doctype/ward/ward.js
@@ -1,0 +1,8 @@
+// Copyright (c) 2025, Frappe and contributors
+// For license information, please see license.txt
+
+// frappe.ui.form.on("Ward", {
+// 	refresh(frm) {
+
+// 	},
+// });
